Restructured file structure
Reordered all server-side files into a module-based structure to outline the different parts in the project
